+++ Utils/Helpers.hpp
@@ -55,7 +55,7 @@ unsigned int	parseIPv4(const std::string& ipAddress);
 std::string		getDate( void );
 std::string		getDate( void );
 std::string		getContentType(const std::string& target, std::map<std::string, std::string>& mimeTypes);
-long			fileLength(std::string& path);
+ssize_t			fileLength(std::string& path);
 std::string		_toString(int num);
 std::string		_toString(long num);
 std::string		_toString(unsigned long num);
